issue #193: publish bigann and gist1m as "custom" datasets on GCS (#194)
## Summary

- New `StandardDatasetPublisher` tool stages BIGANN / GIST1M in the flat
layout VectorBench "custom" mode expects, generates the descriptor JSON,
and optionally runs `push_dataset_gcs.sh` to upload to `gs://`.
Motivation: serve the standard datasets from `gs://herddb-datasets` so
cloud runs stop depending on the TEXMEX FTP mirror.
- Loader gains a `format` field in the descriptor (`fvecs` / `bvecs`,
defaulting to `fvecs`) so `--dataset custom` can consume BIGANN's native
BVECS files. `totalVectors`, `numQueries`, `Config.numRows` and
`Config.resumeFrom` are widened to `long` — BIGANN 1 B already stresses
`int` and 20 B-class datasets are coming.
- BIGANN's nested `bigann_gnd/idx_1000000000.ivecs` is flattened to
`bigann_groundtruth.ivecs` in the staging dir so it lives next to the
descriptor.

Closes #193.

## Test plan

- [x] `mvn -pl vector-testings -B checkstyle:check apache-rat:check
spotbugs:check install -DskipTests -Pci` — green
- [x] `mvn -pl vector-testings test` — 46 tests pass (includes the new
`StandardDatasetPublisherTest` covering FVECS + BVECS round-trip and
BIGANN ground-truth flattening, plus a back-compat test for descriptors
without `format`)
- [ ] Manual end-to-end (not run here): stage GIST1M and run `./run.sh
--dataset custom --dataset-url
gs://herddb-datasets/gist1m/gist1m_descriptor.json -n 1000000
--checkpoint` to confirm ingestion + search
- [ ] Manual end-to-end (not run here): same against BIGANN with a small
`-n` to exercise the BVECS custom-mode code path

The actual ~128 GB BIGANN upload to `gs://herddb-datasets/` is intended
to be run out-of-band by the maintainer once this tooling lands.

+++ vector-testings/src/main/java/herddb/vectortesting/VectorBench.java
@@ -151,7 +151,7 @@ public class VectorBench {
             out.info("Ground truth not available: " + e.getMessage());
         }
 
-        int actualRows = config.numRows;
+        long actualRows = config.numRows;
 
         // Phase 2: Drop table (if requested)
         if (config.dropTable) {
@@ -189,7 +189,7 @@ public class VectorBench {
 
         // Phase 4: Ingestion
         if (!config.skipIngest) {
-            int toIngest = actualRows - config.resumeFrom;
+            long toIngest = actualRows - config.resumeFrom;
             if (toIngest <= 0) {
                 out.info("resumeFrom (" + config.resumeFrom + ") >= rows (" + actualRows + "), nothing to ingest.");
             } else {
@@ -219,7 +219,7 @@ public class VectorBench {
 
             // Progress display thread runs during the entire ingestion
             AtomicBoolean ingestDone = new AtomicBoolean(false);
-            final int totalRowsTarget = config.numRows;
+            final long totalRowsTarget = config.numRows;
             Thread progressThread = new Thread(() -> {
                 Runtime rt = Runtime.getRuntime();
                 while (!ingestDone.get()) {
@@ -233,7 +233,7 @@ public class VectorBench {
 
                     LinkedHashMap<String, Object> fields = new LinkedHashMap<>();
                     fields.put("rows", rowsIngested);
-                    fields.put("total", (long) totalRowsTarget);
+                    fields.put("total", totalRowsTarget);
                     fields.put("ops_per_sec", opsPerSec);
                     fields.put("eta_s", etaSecs);
                     fields.put("commits", commitsTotal.get());
@@ -605,7 +605,7 @@ public class VectorBench {
 
         LinkedHashMap<String, Object> summaryFields = new LinkedHashMap<>();
         summaryFields.put("dataset", config.dataset.name());
-        summaryFields.put("rows", (long) config.numRows);
+        summaryFields.put("rows", config.numRows);
         summaryFields.put("similarity", config.effectiveSimilarity());
         summaryFields.put("total_wall_time_s", round1(totalWallSecs));
         out.summary(summaryFields);
